transactions change JS into TS
